refactor(auth): [GREEN] implement default password policy
